+++ src/app/components/navbar/navbar.vue
@@ -31,7 +31,10 @@
 <script>
 "use strict";
 
+const app = require('electron').remote;
+const dialog = app.dialog;
 const ipcRenderer = require('electron').ipcRenderer;
+
 const EventHandler = require('../../event_handler');
 const {
     SchedulerAlert
@@ -66,6 +69,23 @@ module.exports = {
                     break;
                 case "stopSuite":
                     this.$store.dispatch("stopSuite", this.$store.state.tasks.selectedSuite);
+                    break;
+                case "importSuite":
+                    if (this.$store.getters.canAddSuite) {
+                        dialog.showOpenDialog({
+                            filters: [{
+                                name: 'json',
+                                extensions: ['json']
+                            }]
+                        }).then((dialogResult) => {
+                            if (dialogResult.filePaths && dialogResult.filePaths[0]) {
+                                const filename = dialogResult.filePaths[0];
+                                this.$store.dispatch("importSuite", filename).catch((err) => {
+                                    console.warn(err);
+                                });
+                            }
+                        });
+                    }
                     break;
                 case "scheduleSuite":
                     new SchedulerAlert("Schedule Task Execution").toggle().then((res) => {