+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * Quit option in context menu
+* Import/Export suites
 * Modal alerts styles changed for better consistency
 * Dependencies updated, electron v7